Fix setting SRIOMAINTX_MAC_CTRL on CN63XX pass 1.x. Implement new reset sequence sent by Mike Bloom to octeon-ii_debug. Here is the email:
Post-Silicon Validation has successfully implemented and tested a change to SRIO initialization software that prevents a DID Timeout resulting from a QLM Reset issue (which itself is the subject of an upcoming Errata).  These DID Timeouts occurred on approximately 25-50% of the attempts to run an SRIO DMA engine test.  On each attempt, the failure occurred at the beginning of the test when it initializes an SRIO host mode port at a 3.125GHz data rate on CN6300 Pass 2.1 devices.

The software change is described as follows:
1.       The change is made to cvmx_srio_initialize in executive/cvmx-srio.c .
2.       When the link is disabled to make configuration changes, program the following bits in port_0_ctl.s as follows:
a.       o_enable =0
b.      i_enable = 0
c.       prt_lock = 1
d.      The disable bit’s powerup value is 0.  This can be set to zero again or left alone, but it must not be set.
3.       Program any SRIO CSRs to effect any desired configuration changes.
4.       When the link is to be enabled:
a.       Bring the link down, then up, by writing to the SRIO port’s PORT_0_CTL2 CSR.
        i.      Read CVMX_SRIO_MAINTX_PORT_0_CTL2 into port_0_ctl2.u32.
        ii.      Write port_0_ctl2.u32 back to CVMX_SRIO_MAINTX_PORT_0_CTL2.
b.      Program the following bits in port_0_ctl.s as follows:
        i.      o_enable =1
        ii.      i_enable = 1
        iii.      prt_lock = 0
        iv.      Optionally, disable=0
        v.      Write port_0_ctl.u32 to CVMX_SRIOMAINTX_PORT_0_CTL

git-svn-id: http://casvn/svn/octeon/bdk/trunk@57697 c83883e8-9f28-0410-baf1-f2025811e16a