Add commands to control the debugger's text gui.
Commands have been added to get all information that is displayed
in the GUI. A new "gui" command allows turning the display on and
off. In addition, "-nogui" can disable the display before the 
debugger starts. This allows output from the debugger to be clean
enough to be display without terminal support. This should also
allow SciTE to control it in the future.

git-svn-id: http://casvn/svn/octeon/bdk/trunk@58463 c83883e8-9f28-0410-baf1-f2025811e16a