Redo ethernet FCS/CRC handling.
During bringup it ca nbe useful to have the ethernet FCS available for incomming packets. Change
the bdk_if_receive setup for all ports to leave the FCS on the port. The received packet is 4 bytes
longer than normaly, but the TCP stack ignores padding anyway. The behavior on the various
interfaces is expected to be:

MGMT: FCS present and counted on TX and RX
SGMII: FCS present and counted on TX and RX
XAUI: FCS present and counted on TX and RX
ILK: FCS present and counted on TX and RX
DPI: No FCS.
LOOP: No FCS.
SRIO: No FCS.

git-svn-id: http://casvn/svn/octeon/bdk/trunk@58739 c83883e8-9f28-0410-baf1-f2025811e16a