Update how TX packets are generated in traffic-gen to use FPA buffers.
Previously traffic-gen used a single buffer that was malloc in a single chunk
to hold TX buffers. This meant that the packet could not be freed by the PKO
hardware, which caused a race condition between PKO sending and traffic-gen
doing the free. This change uses FPA buffers for the packet so the last
traffic-gen packet can tell PKO to free. This greatly complicates the building
of the TX packet, so bdk_if gained some functions for accessing data in packets.
Building packets is now quite slow, but it isn't done very often.

git-svn-id: http://casvn/svn/octeon/bdk/trunk@58754 c83883e8-9f28-0410-baf1-f2025811e16a