Undo most of my work creating a Lua CSR database. Sometimes I'm an idiot.
Even though a future goal is to remove the CSR database from the target, this
doesn't mean I can't use the code on the host. For some idiotic reason I got 
it in my head that removing the C based CSR database from the target must also
remove it from the host. This is clearly false. The host can continue to use the
existing C based database. As an added benefit the database continues to exist 
in a format useable on the target for chips with more space.

git-svn-id: http://casvn/svn/octeon/bdk/trunk@62029 c83883e8-9f28-0410-baf1-f2025811e16a