Disable frame alignment if using preamble. This allows the link to reach full rate for odd length packets. For example, without this SGMII is slower than RGMII for 65 byte packets.
The original email explaining this from Dan Katz:

Second, the reset value was chosen to be conservative. Since Octeon
supports a wiiiiiiiide range of frame types, not all communication
requires packets with preambles. In fact, we often get questions from
customers who are running across back-planes and are trying to squeeze
more performance so they kill the preamble. Since this feature eats
into the preamble (the preamble becomes variable and disposable), those
packets would have been corrupted. So we chose a mode that would not
break but may not be optimal in the common case.

Third, in terms of picking the default value, I guess there are a
couple of ways to look at it. Maybe its better to fail with a packet
error that's a more hard error and will get people's attention faster
than failing by having a subtle 1.2% performance problem. I think a
good rule may be that if the TX side is pre-pending the PREAMBLE, then
the bit should be whacked.

Something like:

GMX#_TX@_SGMII_CTL[ALIGN] = !GMX#_TX@_APPEND[PREAMBLE]

I think that's a pretty good value to go with.

git-svn-id: http://casvn/svn/octeon/bdk/trunk@65538 c83883e8-9f28-0410-baf1-f2025811e16a