Per request from Frank Corcoran, request a RX equalization evaluation after any TX QLM parameter is changed. Also add a menu item to do it manually.
Original email:

Looking at the link tuning controls in the current BDK it appears that
you have controls for setting the three TX taps in the transmitter.  We
need to add a control to request an RX evaluation in order to update
the CTLE/DFE after the TX taps have been modified.  This should
probably done as a single command as setting the TX taps should always
be followed by an RX evaluation request.
 
The receiver evaluation can be performed with the following sequence;
  
  GSERX_BR_RXX_CTL.RXT_SWM=1
  GSERX_BR_RXX_EER.RXT_EER=1
  Wait for GSERX_BR_RXX_EER.RXT_ESV=1
  GSERX_BR_RXX_CTL.RXT_SWM=0
   
The GSERX_BR_EER register also contains the figure of merit value determined
in the receiver evaluation as well as the directional update fields for each
of the TX taps.  Printing out this information after the receiver evaluation
would be very helpful in determining the optimal values for a given link.

git-svn-id: http://casvn/svn/octeon/bdk/branches/thunder@112176 c83883e8-9f28-0410-baf1-f2025811e16a