[chainboot] Final commit for adding 512kB FAT filesystem.
This commit adds a 512kB FAT filesystem image to the flash. The boot
process changed to include a stage 0 chainloader. The chainloader loads
the BDK image (stage 1) from flash. At this point the boot process is
the same as the normal-boot-crb-2s boot process.

The build creates TWO images:
 - chainboot.bin
   Regular image, includes chainloader, BDK, diagnostics AND FAT filesystem.
 - chainboot-nonfat.bin
   Same as chainboot.bin but DOES NOT include FAT filesystem.
   Use this image if you want to flash a board and preserve the existing
   filesystem data on the flash.

The layout of the flash changed to accomodate the changes:
0x000000: 64KB unused by Thunder
0x010000: 64KB of CVM_CLIB header
0x020000: 192KB of chainloader image (non-trusted)
0x050000: 192KB of chainloader image (trusted)
0x080000: 512k of BDK image
0x100000: 1.5MB of diagnostics
0x280000: 512k of FAT filesystem image (chainboot.bin only)
0x300000: Free space until 0x400000
0x400000: ATF...

Note that the BDK image size can be up to 512k and the diagnostics size
can be up to 1.5M. The FAT filesystem is only used to store
configuration data at this time.

Eventually, as development continues, boot images (BDK, diagnostics,
ATF, UEFI) will transition into the FAT filesystem.

git-svn-id: http://casvn/svn/octeon/bdk/branches/thunder@120337 c83883e8-9f28-0410-baf1-f2025811e16a