Deskew training cleanup, and add "nibble error" detection.
It is considered a deskew nibble error when the deskew training values
for the upper 4-bits or lower 4-bits of a byte differ by more than 33.
Should we encounter nibble errors for specific RAWCARD types A or B,
we will issue a reset to the bit deskew training and start training over
from the beginning.



git-svn-id: http://casvn/svn/octeon/bdk/branches/thunder@121475 c83883e8-9f28-0410-baf1-f2025811e16a