Change the SATA init sequence for drive detection to be closer to notes from Andrew Shinn about RTL simulations. I see no change in behavior with these changes. The original email:
Hi Chandra,
 
 I reviewed your initialization sequence (from email) and also the
BDK source code.  As you mentioned on the call, I do see the
difference in when SCTL.DET is written to 1.  In the BDK, it is
written before CMD and in the SDK, it is written afterwards.
Ends up that in both cases, SCTL.DET is ignored by the host 
controller:
  
  ·         BDK:  writing a 1 to DET is blocked when SUD = 0

  ·         SDK:  writing a 1 to DET is blocked when ST = 1

   
Here are the differences between your SDK sequence and our simulation environment:
    
    ·         SDK

    1.       Write CMD.FRE=1, CMD.SUD=1, CMD.POD=1, CMD.ST=1

    2.       Write SCTL.DET=1

    ·         Simulation environment

    1.       Write CMD.FRE=1, CMD.SUD=1, CMD.POD=1

    2.       1us delay

    3.       Write SCTL.DET=0x1 (0 -> 1 transition resets the port)

    4.       1ms delay

    5.       Write SCTL.DET=0x0 (1 -> 0 transition re-starts the OOB initialization sequence)

    6.       Poll SSTS.DET

    7.       Write CMD.ST=1

     
So I would try the BDK’s sequence (even though SCTL.DET is blocked)
and then the simulation environment’s sequence.  If these two
changes don’t work, then potential issues with the clock source
needs to be investigated.