Allow the DDR_ALT_REFCLK configuration option to force use of the alternate refclk speed specification for all DIMM types and DDR speeds.
Add a Setup Menu item to show and change the DDR_ALT_REFCLK setting.

When the DDR_ALT_REFCLK setting does not always force an alternate refclk,
if the other DRAM configuration conditions do require use of alternate refclk,
the code will still attempt initially to configure it and will use it if
it is present, otherwise it will gracefully return to the standard 50MHz.