Change the SSO lookup table to be dynamically allocated.
With the possibility of 4 nodes and 1024 channels, the SSO channel to
handle lookup table was quite large. It is not used at all on CN88XX
and CN81XX. One CN83XX, only one node is used. We now only allocate
a table for a node when it is actually needed.