Fix TWSI write issue causing XLAUI to not work on the CRB-1S
Due to a typo in a scanf() line, the TWSI write data was always zero
if it was specified in hex. This patch changes the foramt string to
allow hex to work. For added safety, the PHY is given 100ms to complete
reset before the BDK continues with MDIO init.