[cortina] Don't print "Update successful" in low level routine.
The update process will be "successful" even though the write enable
jumper on the SPI flash may not be set. In this case we do not want
to print the "Update successful" message. We check the firmware date
in the calling function to determine if the image actually made it onto
the flash.