Move setting of the DRAM SPD addresses out of the common DDR3 and DDR4 config
UEFI uses the SPD addresses to determine if a DIMM slot exists even if the DIMM
is not present. This is needed so it can properly populate the SMBIOS tables. The
default SPD addresses in the DDR* configs was causing UEFI to falsely report
missing DIMMs on boards with 4 DIMM slots instead of 8. The inheretance property
of the BDK config system caused 4 slot boards to pull in the defaults. With this
change, all ODM boards will need to specify the SPD address for their slots as
no default is given.