Create a UAA driver for use with uarts 2 and above
CN81XX and CN83XX have four uarts instead of just two. This
driver connects the uarts to the console system. The second
node of a two node system can now be used as a BDK console.
Note that although the UAA device is hooked up, the GPIO
configuration is per board and not configured by this code.