Change SERDES lane tuning to not override values when defaults are used
The tuning function used to read the default values and program these as
fixed if the user didn't supply a value. This breaks for PCIe where the
defaults are readable and we get zero. Now the code only sets enables
for the values specified. One change is that you must program the pre and
post settings to get either to apply.