Device tree fixup for Linux
The SDK wants to push all board specific infirmation into the BDK so
ATF, UEFI, and Uboot can be generic. To support this, the BDK will
load a device tree from FAT and pass it to ATF. The device tree may
be generic for Thunder, which means all possible SOC configs are
present. The BDK must prune out unused entries before passing the
tree to ATF.