Change how KR training is tracked for BGX interfaces
The BDK used to use bits in GSERx_RX_TXDIR_CTRL_1 to track
if a BGX port should use KR training. We have to track this
in hardware so it survives jumping between binaries. The bits
in GSERx_RX_TXDIR_CTRL_1 were a poor choice as they corrupt
the direction hints from RX equalization. To fix this, the
BDK nows uses bits in GSERX_SCRATCH to store the KR
state. To simplify the QLM code across chips, functions have
been added to bdk-qlm-common.c to abstract this.