Use recursive lock for console output instead of spinlock
A deadlock could occur between the spinlock in the console output
routines and the low level uart recursive lock. The problem was
that the low level lock could yield to another thread on the same
core, which calls printf. The second printf would then stall on
the spinlock and not yield.