Update the global power throttle calculation based on new information
From David Carlson:
The fuses are a straight percent with hardware doing the multiplication. For
the csr, you have to multiply(the intent was we'd use the fuses). You would
do 48*165*4*0.65 = 20592.
The formula hardware uses is: 165*4*percent*num_cores/100

165 is MAXPOW from CVM_POWER_EL1. It's a different value between thunder and
octeon but otherwise the same for all chips in that process/architecture
family. 165*4 is the maximum power in mA/Ghz which is what's used
internally. So the final number you supply global throttling will roughly
be mA/Ghz(of rclk).