Add function to initialize fuses before any uart output
We are starting to need the ability to program various fuses
early in boot to change hardware defaults before parts of the
chip come out of reset. Soft blowing fuses normally requires
a soft reset to get the various values pulled into the correct
locations in the chip. Do this before uart output so the user
can't see it happening.

Currently this function only changes the default throttle level
of the cores to 66%.