Change fuse blowing routines to not need the GTI clock
We need to be able to soft blow fuses in early init code, and
on secondary nodes before CCPI. In these setups, the GTI
timer isn't running yet. Make the spin loops in the fuse
program infinite. This is normally bad, but the fuse hardware
has never shown the tendancy to hang.