[Fuse] Revert all programming of fuses during boot (Part 1 of 11)
Between marketing changing SKUs four times in a month,
and OPs lying about what fuses are blown in production,
it is evident that the whole PNAME and throttle fuse
stadegy will not work. This reverts a whole series of
commits related to this. Do not attempt anything similar
in the future without a detailed spec signed off by all
parties.