Add a device tree config item for QLM insertion loss
Some of the GSER settings need to know the insertion
loss for a QLM to be programmed correctly. This change
allows customers to supply the insertion loss in dB
as a configuration item. If not specified, the default
assumes greater than 10 dB, which was the setup before
this was added.