[ccpi] Detect when CN88XX pass 1.0 hardware finishes training too early and skip training
Due to an errata, CN88XX pass 1.0 is not capable of restarting training once it completes.
In an already tuned link, such as after warm reset, some lanes may finish training before
software has a chance to run. This causes the normal software training to not work, so
we skip it on these lanes. Pass 1.0 can still fail if one node is done with training and
the other resets, wanting training again. Right now there is no recovery for pass 1.0 in
this mode. Hopefully all the pass 1.0 systems are gone my now.