[t81] Add Bruce Haley's updated Micrel PHY code
Email:
 
I tested the speeds/duplex combinations for the RGMII port,
and the existing code to get link status information for
Kendin parts didn’t cover 10Mbs, or half-duplex.  The 8104’s
Micrel PHY (which reports as a Kendin) has a PHY Control
register at address 0x1F that seems to report the correct
speeds and duplex status.  Attached is a diff of my changes
to support the Micrel part – not sure how they would affect
BDK support for any existing Kendin parts, so I’m sending
you the code.  Feel free to use or discard as you see fit.

Chad's note: This is the same PHY that the BDK code was
written for, Bruce's code just covers it better. This code
should be back ported to CN70XX sometime.