[xaui] Properly detect link down status for XAUI
This is a slight modification of a patch from Bruce Haley.
His email:

It looks like adding a check for “loc_fault” in the
if_link_get_xaui() function fixes the issue.  The attached
patch has it defined for RXAUI as well, based on the HRM
description of the bit, but it doesn’t appear as though
the RXAUI module needs it – RXAUI drops link just fine
when the cable is pulled without it.

Signed-off-by: Chad Reese <kreese@cavium.com>