[pcie] Set SLIX_M2S_MACX_CTL[waitl_com] for more strict ordering
Testing at Alibaba with the Ali Flash cards have shown issues with
the default weak ordering for PCIe. Setting SLIX_M2S_MACX_CTL[waitl_com]
fixes the issue. This is not as big of a performance impact as WAIT_COM,
so it is ok to have it on all the time.

I was not CCed on the email thread about this, so I don't know the details.