[qlm] Remove code skipping QLM setup if we're already in the correct mode
Early in t88, it was common for QLM init to be called numerous times
due to how images loaded. This is not the case anymore, so I believe
the mode check causes more harm than good. This change has risk. The
existing code didn't handle t81 pcie correctly. Rather than try and
fix it, I think it is safe to remove.