[dram] Fix possible skip of ECC update in memory owned by the BDK
There is a region of Memory between 3/4 of the size of L2 and the size of L2
that may be written to DRAM before DRAM is initialized. Although this region
is unused and doesn't matter, this could cause the ECC bits for these cache
lines to not be updated. This seems to happen most often on ODM boards (foxconn)
which have extra processing in the BDK before DRAM is initialized. This causes
more L2 evicts.