Don't use compiler builtins for compare and swap
CN88XX pass 1.x has an errata where compare and swap
can't be used. GCC 5.x uses compare and swap for
builtin atomics, with no way to disable it on the
command line. Use conditional checks for pass 1.x
and inline assembly to use CAS on chips that
support it, falling back to load/store exclusive
for pass 1.x.

The BDK can now be compiled with both GCC 4.x and GCC 5.x.