[DRAM] Use BDK_TRACE for waiting-for-cores message
In bdk-dram-test(), use BDK_TRACE to output the waiting-for-cores message
when a test is taking a long time to run. The message is not very useful
on CN81XX with only 4 cores, where the Random XOR tests will need a lot of
time to complete. Testing and reporting the condition will still be
available for debugging, as instances of cores going "bad" can lead to
tests getting hung for no apparent reason.