[ccpi] Fix bug where CCPI tuning parameters would default to zero
Due to how image handoff works in the BDK, the function __bdk_init_ccpi_early()
is only called in the first image and not following images. This is the expected
and correct behavior. The bug was that the CCPI lane state was initialized in
this function instead of __bdk_init_ccpi_connection(). This caused the defaults
for CCPI lane tuning to be lost.