[sata] Consolidate QLM+SATA init code into a single common file
The duplicate SATA init code for each chip had gotten out of sync.
Conolidate the details of SATA init into one common QLM file, calling
it for the various chips. As part of this, fix a bug where the SATA
speed was lost during controller reset. This mean that all systems
were using SATA3 regardless of the configured speed. Another artifact
of the bug was that SATA DLM/QLM always looked disabled in the SERDES
menu.