Hack 81xx test script to get it working again. Needs rework
The serial script handling has changed since the EBB8104
network test script was written. This is a hack to get
this test running again so I can check for regressions
with CN80XX changes. The board name and connection are
all hard coded. Needs to be fixed in the future.