[sata] Take SATA PCS out of reset before the GSER PHY
Email from Scott Mcllhenny:

Let’s try this.  In the BDK code Move Step 9: Clear the GSER(0..13)_SATA_LANE_RST[Ln_RST] = 0
-before- Step 4: Take the PHY out of reset:  Write GSER(0..13)_PHY_CTL[PHY_RESET]=0.

What I see in the simulations (waves attached) is taking the SATA_PCS out of reset *before* the PHY
keeps the PHY in the P0 power state as it initializes (which we want).  The SATA_PCS also holds the resets to the SATA
MAC until the PHY signals that it is “Ready” via the MGMT_ACK signals, but since we are programming the
PLLs in the PHY the SATA MAC reset is released too early.  So we’d prefer to hold the SATA MAC in reset until
this PHY signals that it is “Ready” again after the PLL Dividers have been programmed and digested. This is much
cleaner and avoids the PHY P0->P2->P0 power state transitions that I see when the SATA PCS is released
from reset after the PHY.