[net,tso] Change how packet verification CRC32 is calculated
With full offload of TCP and UDP checksums, the packet changes
after transmit such that RX see different header checksums. Modify
the CRC32 in the packet to skip the TCP or UDP checksum, so the
CRC doesn't change. This make packet validation not the same as
octeon, or previous BDKs. A mix of old and new BDKs will get
validation errors.