[twsi] Implement bus recovery after a TWSI timeout
Email 1:
Please review the patch which I believe should fix the Foxconn's
"Timeout waiting for operation to complete" issue.

The patch is made against the 1.23.7 thunderx bdk release.

Please also note that we were not able to reproduce the issue locally at
Cavium and have not actually verified it at Foxconn's.

However, the patch is a BDK port of the original Linux recovery
procedure which has been proven to work.

Regards,
Dmitry

Email 2:
I added a 5ms delay before subsequent retry for the case when one of the
arbitration lost states is recognized.

The idea is that we know that the TWSI hardware is switched to the slave
mode and expects STOP condition on the bus when it recognizes an
arbitration lost.

If we do an immediate retry, we are likely end up with the "timeout
waiting" error due to the known TWSI bug.

So, we do a 5ms delay before a subsequent retry which is more than
enough for normal I2C transfer to end and STOP condition be issued.

Otherwise, the recovery procedure will be taken.

Regards,
Dmitry