[qlm] Add support for SGMII with a 100Mhz ref clock
Email:
Hi Chad,

We got an update from MA that we can support SGMII in 100MHz ref clk with the below settings.
I have modified the BDK to try this on a CN81xx NAS board. Can you please look at the changes and commit it to the main branch? Thanks!

...

Regards,
Srikanth