[bgx] Check for bit errors during BGX link init
For a BGX link to be considered up, require that it
have block lock during the entire up sequence. Also
do a short bit error check to stop link up on false
link bounces. This is a backport of work from CN73XX
for Alibaba 73xx Liquide I/O issue. The changes are
generically applyable.