[t83] Give BDK 32MB of memory under asim
Memory requirements for NIC and PKI/PKO seem to have ballooned
out of control so that the BDK doesn't run with 16MB of memory.
Increase it to 32MB in asim. We'll need to revisit this when we
get the chip. it should be possible to run in cache, which is
8MB.