[dbg] Add define to enable / disable the checking of a magic register dump signal
The threading library can call bdk_dbg_check_magic() to check for a magic system
request, a break signal on uart0, to cause cores to create a register dump. Add
a define to enable or disable this, disabled by default. This is disabled because
some field systems are getting random data aborts on the read of the UAA register.