[ccpi] Send 1/3 of the number of messages during message check
Now that we perform RX equalization between each message during
message check, the time to perform the full check is long. Increment
message counters by 3 so we send 1/3 the number of messages. This
allows more bits in the training message to be checked and for the
whole sequence to run faster.