[t83] Rewrite how PKO tracks DQ depth
PKO accesses packet data after descriptor words are
removed from a DQ. This means using watermarks or
PKO's memory flow control will show the DQ as empty
even though the packet data is in use. The BDK needs
to know when the packet is finished, so we now use
memory atomics to track DQ depth. This requires two
atomics per packet. A fetch and add on submit, and
PKO atomic decrement on finish.